% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collaborators.R
\name{add_collaborator}
\alias{add_collaborator}
\title{Add a Collaborator to a Repository}
\usage{
add_collaborator(
  owner,
  repo,
  username,
  permission = c("push", "pull", "admin")
)
}
\arguments{
\item{owner}{Name of the repository owner.}

\item{repo}{Name of the repository.}

\item{username}{Name of the user.}

\item{permission}{The permission to grant the collaborator. Can be one of:
`"pull"`, `"push"`, or `"admin"`. Default: `"push"`}
}
\description{
Invite a user to a repository as a collaborator.
}
\section{Permission}{


The permission to grant the collaborator.
**Only valid on organization-owned repositories.** Can be one of:
- `"pull"`
    - can pull, but not push to or administer this repository
- `"push"`
    - can pull and push, but not administer this repository.
- `"admin"`
    - can pull, push and administer this repository.
}

\references{
<https://developer.github.com/v3/repos/collaborators/#add-user-as-a-collaborator>
}
