% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commits.R
\name{create_url_commit_diff}
\alias{create_url_commit_diff}
\alias{browse_commit_diff}
\title{View Web Interface Showing Commit Differences}
\usage{
create_url_commit_diff(user, repo, start, end)

browse_commit_diff(user, repo, start, end)
}
\arguments{
\item{user}{Name of the User Account}

\item{repo}{Name of the Repository}

\item{start}{Start commit, tag, date, or branch}

\item{end}{End commit, tag, date, or branch}
}
\value{
A URL of the form:

`https://github.com/<user>/<repo>/compare/<start>...<end>`
}
\description{
Generates a URL to GitHub's web-based commit comparison interface and
opens it if session is interactive.
}
\details{
The gist of possible comparisons is given in the next table.

| Comparison Type   |  start              |     end                                   |
|:------------------|:-------------------:|:-----------------------------------------:|
| Forks             | `master`            | `octo-org:master`                         |
| Tags              | `v2.2.0`            | `octocat:v2.3.3`                          |
| Branches          | `master`            | `octocat:an-example-comparison-for-docs`  |
| Time              | `master@{2weeks}`   | `master`                                  |
| Time              | `master@{MM-DD-YY}` | `master`                                  |
| Commits           | `c3a414e`           | `faf7c6f`                                 |
| Commits           | `96d29b7^^^^^`      | `96d29b7`                                 |
}
\examples{

browse_commit_diff("coatless", "ghapi3", start = "ab469f8", end = "f15f06e")

}
\references{
<https://help.github.com/en/articles/comparing-commits-across-time>
}
