% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{create_tag}
\alias{create_tag}
\title{Create a Tag}
\usage{
create_tag(owner, repo, tag, message, object, type)
}
\arguments{
\item{owner}{Name of the person or organization who owns the repository.}

\item{repo}{Name of the repository.}

\item{tag}{The tag's name. This is typically a version (e.g., "v0.0.1").}

\item{message}{The tag message.}

\item{object}{The SHA of the git object this is tagging.}

\item{type}{The type of the object we're tagging. Normally this is a commit but it can also be a tree or a blob.}
}
\description{
Tag a specific portion of code
}
\references{
<https://developer.github.com/v3/git/tags/#create-a-tag-object>
}
