% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{delete_gh}
\alias{delete_gh}
\title{Wrapper for GitHub client's function}
\usage{
delete_gh(...)
}
\arguments{
\item{...}{Additional parameters}
}
\description{
Intermediary wrapper into the `gh` packages `gh` function.
}
\details{
Wrapper function to `gh` for more control over package options
}
