% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collaborators.R
\name{get_collaborator_list}
\alias{get_collaborator_list}
\title{List Collaborators}
\usage{
get_collaborator_list(owner, repo, affiliation = c("all", "outside", "direct"))
}
\arguments{
\item{owner}{Name of the repository owner.}

\item{repo}{Name of the repository.}

\item{affiliation}{The affiliation level of collaborators. Can be one of:
`"outside"`, `"direct"`, or `"all"`. Default: `"all"`}
}
\description{
Obtain the names of all collaborators
}
\section{Affiliation}{


Filter collaborators returned by their affiliation. Can be one of:
- `"outside"`
    - All outside collaborators of an organization-owned repository.
- `"direct"`
    - All collaborators with permissions to an organization-owned repository,
     regardless of organization membership status.
- `"all"`
    - All outside collaborators of an organization-owned repository.
}

\references{
<https://developer.github.com/v3/repos/collaborators/#list-collaborators>
}
