% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repos.R
\name{get_contributors_list}
\alias{get_contributors_list}
\title{Retrieve list of contributors for a repo}
\usage{
get_contributors_list(owner, repo, anon = TRUE)
}
\arguments{
\item{owner}{Name of the person or organization who owns the repository.}

\item{repo}{Name of the repository.}

\item{anon}{Either `TRUE` to include anonymous contributors in results or
`FALSE` to not include anonymous contributors. Default:
`TRUE`.}
}
\description{
Obtains a list of people who have contributed code or text to a
repository
}
\references{
<https://developer.github.com/v3/repos/#list-contributors>
}
