% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_emoji_list}
\alias{get_emoji_list}
\title{List of Emoji Code and Image URLs}
\usage{
get_emoji_list()
}
\description{
Lists all the emojis available to use on GitHub.
}
\references{
<https://developer.github.com/v3/emojis/#emojis>
}
