% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organization.R
\name{get_org_invitations}
\alias{get_org_invitations}
\title{List Organization Invitations}
\usage{
get_org_invitations(org)
}
\arguments{
\item{org}{Name of the organization}
}
\description{
Retrieves all pending organization invitations
}
\references{
GitHub API Documentation reference for public membership check
<https://developer.github.com/v3/orgs/members/#list-pending-organization-invitations>
}
