% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organization.R
\name{get_org_members_list}
\alias{get_org_members_list}
\title{Retrieve Membership list in an Organization}
\usage{
get_org_members_list(
  org,
  filter = c("all", "2fa_disabled"),
  role = c("all", "admin", "member")
)
}
\arguments{
\item{org}{Name of the organization}

\item{filter}{Filter members returned in the list. Can be one of: `"2fa_disabled"`
or `"all"`. Default: `"all"`}

\item{role}{Filter members returned by their role. Can be one of: `"all"`,
`"admin"`, or `"member"`. Default: `"all"`}
}
\description{
Retrieves the usernames of people in an Organization
}
\section{Filter}{


- `"2fa_disabled"`
    - All members the authenticated user can see.
- `"all"`
    - Members without two-factor authentication enabled. Available for organization owners.
}

\section{Members List}{


- `"all"`
    - All members of the organization, regardless of role.
- `"admin"`
    - Organization owners.
- `"member"`
    - Non-owner organization members.
}

\references{
<https://developer.github.com/v3/orgs/members/#members-list>
}
