% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repos.R
\name{get_org_repos_list}
\alias{get_org_repos_list}
\title{List Organization Repositories}
\usage{
get_org_repos_list(org, type = "all")
}
\arguments{
\item{org}{Name of the GitHub Organization to retrieve repos for.}

\item{type}{Can be one of `"all"`, `"public"`, `"private"`, `"forks"`,
`"sources"`, `"member"`. Default: `"all"`.}
}
\description{
Generates a list of organization repositories
}
\references{
<https://developer.github.com/v3/repos/#list-organization-repositories>
}
\seealso{
[`get_self_repos_list`], [`get_user_repos_list`]
}
