% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teams.R
\name{get_org_team_invitations}
\alias{get_org_team_invitations}
\title{List pending team invitations}
\usage{
get_org_team_invitations(id)
}
\arguments{
\item{id}{Team ID Number.}
}
\description{
Retrieves a list of all the pending invitations to a given team
}
\references{
<https://developer.github.com/v3/teams/members/#list-pending-team-invitations>
}
