% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repos.R
\name{get_self_repos_list}
\alias{get_self_repos_list}
\title{List Self Repos}
\usage{
get_self_repos_list(
  visibility = "all",
  affiliation = "owner,collaborator,organization_member",
  type = "all",
  sort = "full_name",
  direction = "desc"
)
}
\arguments{
\item{visibility}{Can be one of `"all"`, `"public"`, or `"private"`. Default: `"all"`}

\item{affiliation}{Comma-separated list of values. Can include: `"owner"`,
`"collaborator"`, `"organization_member"`.
Default: `"owner,collaborator,organization_member"`.}

\item{type}{Can be one of `"all"`, `"public"`, or `"private"`.
Default: `"all"`. Will cause a `422`` error if used in the
same request as `visibility` or `affiliation`.}

\item{sort}{Can be one of `"created"`, `"updated"`, `"pushed"`, `"full_name"`.
Default: `"full_name"`.}

\item{direction}{Can be one of `"asc"` or `"desc"`. Default: `"desc"`}
}
\description{
Retrieve information on the repositories created or held by yourself.
}
\references{
<https://developer.github.com/v3/repos/#list-your-repositories>
}
\seealso{
[`get_user_repos_list`], [`get_org_repos_list`]
}
