% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repos.R
\name{get_user_repos_list}
\alias{get_user_repos_list}
\title{List User Repositories}
\usage{
get_user_repos_list(
  username,
  type = "owner",
  sort = "full_name",
  direction = "desc"
)
}
\arguments{
\item{username}{Name of the GitHub user to retrieve repos for.}

\item{type}{Can be one of `"all"`, `"owner"`, or `"member"`.
Default: `"owner"`.}

\item{sort}{Can be one of `"created"`, `"updated"`, `"pushed"`, `"full_name"`.
Default: `"full_name"`.}

\item{direction}{Can be one of `"asc"` or `"desc"`. Default: `"desc"`}
}
\description{
Generates a list of user repositories
}
\references{
<https://developer.github.com/v3/repos/#list-user-repositories>
}
\seealso{
[`get_self_repos_list`], [`get_org_repos_list`]
}
