% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghapi-package.R
\docType{package}
\name{ghapi-package}
\alias{ghapi}
\alias{ghapi-package}
\title{ghapi: GitHub v3 API Implementation}
\description{
Provides access to GitHub API v3.0 using the gh package backend.
}
\details{
This package requires users have access to a GitHub Personal Access Token (PAT)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/coatless/ghapi}
  \item Report bugs at \url{https://github.com/coatless/ghapi/issues}
}

}
\author{
\strong{Maintainer}: James Balamuta \email{balamut2@illinois.edu} [copyright holder]

Other contributors:
\itemize{
  \item GitHub Inc. \email{support@github.com} [copyright holder]
}

}
