% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organization.R
\name{create_org_invitation_ghname}
\alias{create_org_invitation_ghname}
\alias{create_org_invitation_email}
\title{Create Organization Invitations}
\usage{
create_org_invitation_ghname(
  org,
  invitee_id,
  role = c("admin", "direct_member", "billing_manager"),
  team_ids = NULL
)

create_org_invitation_email(
  org,
  email,
  role = c("admin", "direct_member", "billing_manager"),
  team_ids = NULL
)
}
\arguments{
\item{org}{Name of the organization}

\item{invitee_id}{GitHub user ID for the person you are inviting.
Not required if you provide email.}

\item{role}{Specify role for new member. Can be one of:}

\item{team_ids}{Specify IDs for the teams you want to invite new members to.}

\item{email}{Email address of the person you are inviting, which can be
an existing GitHub user. Not required if you provide invitee_id.}
}
\description{
Invite multiple people to an organization using their GitHub user ID or their
e-mail address.
}
\details{
This set of API bindings is a WIP and requires additional values.
}
\section{Role}{


Specify `role` for new member. Can be one of:

- `admin`
     - Organization owners with full administrative rights to the
       organization and complete access to all repositories and teams.
- `direct_member`
     - Non-owner organization members with ability to see other members
       and join teams by invitation.
- `billing_manager`
     - Non-owner organization members with ability to manage the billing
       settings of your organization.
}

\references{
GitHub API Documentation reference for public membership check
<https://developer.github.com/v3/orgs/members/#create-organization-invitation>
}
