% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organization.R
\name{add_org_member}
\alias{add_org_member}
\alias{remove_org_member}
\alias{update_org_member_role}
\alias{check_org_public_membership}
\title{Invite, Remove, or Modify User Membership in an Organization}
\usage{
add_org_member(org, username, role = c("member", "admin"))

remove_org_member(org, username)

update_org_member_role(org, username, role = c("member", "admin"))

check_org_public_membership(org, username)
}
\arguments{
\item{org}{Name of the organization.}

\item{username}{Name of the user on GitHub.}

\item{role}{The role to give the user in the organization. Can be one of:
`"admin"` or `"member"`. Default: `"member"`}
}
\description{
Sends an invitation, removes, or modifies a user role within an organization.
}
\details{
To prevent abuse, the authenticated user is limited to 50 organization
invitations per 24 hour period. If the organization is more than one month
old or on a paid plan, the limit is 500 invitations per 24 hour period.
}
\section{Role}{


- `"admin"`
    - The user will become an owner of the organization.
- `"member"`
    - The user will become a non-owner member of the organization.
}

\references{
<https://developer.github.com/v3/orgs/members/#add-or-update-organization-membership>

<https://developer.github.com/v3/orgs/members/#remove-a-member>

<>
}
