% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collaborators.R
\name{remove_collaborator}
\alias{remove_collaborator}
\title{Remove a Collaborator from a Repository}
\usage{
remove_collaborator(owner, repo, username)
}
\arguments{
\item{owner}{Name of the repository owner.}

\item{repo}{Name of the repository.}

\item{username}{Name of the user.}
}
\description{
Delete a repository collaborator.
}
\references{
<https://developer.github.com/v3/repos/collaborators/#remove-user-as-a-collaborator>
}
