% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repos.R
\name{repo_transfer}
\alias{repo_transfer}
\title{Transfer a repository}
\usage{
repo_transfer(owner, repo, new_owner, team_id = NA)
}
\arguments{
\item{owner}{The username or organization name that controls the
repository.}

\item{repo}{The name of the repository to transfer.}

\item{new_owner}{The username or organization name the repository will be
transferred to.}

\item{team_id}{ID of the team or teams to add to the repository.
Teams can only be added to organization-owned repositories.}
}
\description{
A transfer request will need to be accepted by the new owner when
transferring a personal repository to another user. The response will
contain the original owner, and the transfer will continue asynchronously.
For more details on the requirements to transfer personal
and organization-owned repositories, see
[about repository transfers](https://help.github.com/articles/about-repository-transfers/).
}
\references{
<https://developer.github.com/v3/repos/#transfer-a-repository>
}
