% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teams.R
\name{get_org_teams_list}
\alias{get_org_teams_list}
\alias{get_org_team_details}
\title{List teams or retrieve a team's details}
\usage{
get_org_teams_list(org)

get_org_team_details(id)
}
\arguments{
\item{org}{Organization name.}

\item{id}{Team ID Number.}
}
\description{
Generate a list of teams in the organization or
retrieve a specific team's information.
}
\references{
<https://developer.github.com/v3/teams/#list-teams>

<https://developer.github.com/v3/teams/#get-team>
}
