% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teams.R
\name{add_org_team_member}
\alias{add_org_team_member}
\alias{remove_org_team_member}
\title{Add or update team membership}
\usage{
add_org_team_member(id, username, role = c("member", "maintainer"))

remove_org_team_member(id, username)
}
\arguments{
\item{id}{Team ID Number.}

\item{username}{Name of the user on GitHub.}

\item{role}{The role that this user should have in the team. Can be one of:
`"member"` or `"maintainer"`. Default: `"member"`}
}
\description{
Includes a GitHub user as a member of an Organization's team or updates
their membership status.
}
\section{Role}{

The role that this user should have in the team. Can be one of:
- `"member"`
    - a normal member of the team.
- `"maintainer"`
    - a team maintainer. Able to add/remove other team members, promote other
      team members to team maintainer, and edit the team's name and description.
}

\section{Remove Team Member}{

This does not delete the user on GitHub nor does it remove the user from
the GitHub organization. This functionality only suppresses users from
being identified as a member of a team.
}

\references{
<https://developer.github.com/v3/teams/members/#add-or-update-team-membership>

<https://developer.github.com/v3/teams/members/#remove-team-membership>
}
