% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organization.R, R/repos.R
\name{get_org_repos_list}
\alias{get_org_repos_list}
\title{Retrieve Repositories in an Organization}
\usage{
get_org_repos_list(org, type = "all")

get_org_repos_list(org, type = "all")
}
\arguments{
\item{org}{Name of the organization}

\item{type}{The kind of repository to list. Can be one of:
`"all"`, `"public"`, `"private"`, `"forks"`, `"sources"`, `"member"`.
Default: `"all"`}

\item{org}{Name of the GitHub Organization to retrieve repos for.}

\item{type}{Can be one of `"all"`, `"public"`, `"private"`, `"forks"`,
`"sources"`, `"member"`. Default: `"all"`.}
}
\description{
Obtains repositories that belong to an organization according to
different levels.

Generates a list of organization repositories
}
\examples{
get_org_repos_list("stat385uiuc", "all")
}
\references{
<https://developer.github.com/v3/repos/#list-organization-repositories>

<https://developer.github.com/v3/repos/#list-organization-repositories>
}
\seealso{
[`get_self_repos_list`], [`get_user_repos_list`]

[`get_self_repos_list`], [`get_user_repos_list`]
}
