% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghapi3-package.R
\docType{package}
\name{ghapi3-package}
\alias{ghapi3}
\alias{ghapi3-package}
\title{ghapi3: GitHub v3 API Implementation}
\description{
Provides access to select GitHub API v3.0 endpoints using the 'gh' package backend.
}
\details{
This package requires users have access to a GitHub Personal Access Token (PAT)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/coatless-rpkg/ghapi3}
  \item \url{https://r-pkg.thecoatlessprofessor.com/ghapi3/}
  \item Report bugs at \url{https://github.com/coatless-rpkg/ghapi3/issues}
}

}
\author{
\strong{Maintainer}: James Balamuta \email{balamut2@illinois.edu} (\href{https://orcid.org/0000-0003-2826-8458}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item GitHub Inc. \email{support@github.com} [copyright holder]
}

}
\keyword{internal}
